+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Add a new argument to `table` API to allow it to be a super table. ([#159](https://github.com/sdispater/tomlkit/pull/159))
+- Support adding item to `Table` and `Container` with dotted key. ([#160](https://github.com/sdispater/tomlkit/pull/160))
 
 ## [0.8.0] - 2021-12-20